+++ town-api/src/main/java/entity/NotifyUserInfoDO.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
 @Data
 public class NotifyUserInfoDO implements Serializable {
     private Integer id;
-    private Integer userTel;
+    private Long userTel;
     private Integer msgType;
     private byte[] msgCtx;
 }
