+++ town-api/src/main/java/entity/QuestionHandlingInfoDO.java
@@ -9,7 +9,7 @@ public class QuestionHandlingInfoDO implements Serializable {
     private Integer handleId;
     private Integer questionId;
     private Integer handlingType;
-    private Integer handleUserTel;
+    private Long handleUserTel;
     private String handleCtx;
     private Long handleTime;
 }
